Refactor HITL module: split request/response/event classes into dedicated files, optimize imports, and enhance modularity.
