+++ aihub_agent/playground/minimal_workflow/human_in_the_loop_workflow/HumanInTheLoopAgent.py
@@ -1,5 +1,5 @@
 from aihub_lib.nats.events import StartEvent, StopEvent
-from aihub_lib.nats.events.human_in_the_loop.HumanInTheLoop import HumanInTheLoopInput
+from aihub_lib.nats.events.human_in_the_loop import HumanInTheLoopInput
 
 from aihub_agent.agents.Agent import Agent
 from aihub_agent.workflow.decorators.step import step
