+++ aihub_agent/playground/minimal_workflow/human_in_the_loop_workflow/tests/test_HumanInTheLoopAgent.py
@@ -1,12 +1,8 @@
 from aihub_lib.i18n.LocaleString import LocaleString
 from aihub_lib.nats.events import StartEvent
-from aihub_lib.nats.events.human_in_the_loop.HumanInTheLoop import HumanInTheLoopInput
-from aihub_lib.nats.events.human_in_the_loop.request.HumanInTheLoopRequestEvent import (
-    HumanInTheLoopInputRequestEvent,
-)
-from aihub_lib.nats.events.human_in_the_loop.response.HumanInTheLoopResponseEvent import (
-    HumanInTheLoopInputResponseEvent,
-)
+from aihub_lib.nats.events.human_in_the_loop import HumanInTheLoopInput
+from aihub_lib.nats.events.human_in_the_loop.request import HumanInTheLoopInputRequestEvent
+from aihub_lib.nats.events.human_in_the_loop.response import HumanInTheLoopInputResponseEvent
 from aihub_lib.testing.asyncio_utils.bdd import async_test
 from pytest_bdd import given, parsers, scenarios, then, when
