+++ aihub_agent/playground/minimal_workflow/multistep_human_in_the_loop_workflow/events/SecondStepHumanInTheLoop.py
@@ -1,10 +1,6 @@
-from aihub_lib.nats.events.human_in_the_loop.HumanInTheLoop import HumanInTheLoopInput
-from aihub_lib.nats.events.human_in_the_loop.request.HumanInTheLoopRequestEvent import (
-    HumanInTheLoopInputRequestEvent,
-)
-from aihub_lib.nats.events.human_in_the_loop.response.HumanInTheLoopResponseEvent import (
-    HumanInTheLoopInputResponseEvent,
-)
+from aihub_lib.nats.events.human_in_the_loop import HumanInTheLoopInput
+from aihub_lib.nats.events.human_in_the_loop.request import HumanInTheLoopInputRequestEvent
+from aihub_lib.nats.events.human_in_the_loop.response import HumanInTheLoopInputResponseEvent
 
 
 class SecondStepHumanInTheLoopRequestEvent(HumanInTheLoopInputRequestEvent):
