+++ aihub_lib/aihub_lib/nats/events/human_in_the_loop/request/HumanInTheLoopRequestEvent.py
@@ -40,51 +40,3 @@ class HumanInTheLoopRequestEvent(ControlAndDisplayEvent):
             description="The type of HITL interaction: 'input' for free-form text, 'confirmation' for yes/no.",
         ),
     ]
-
-
-class HumanInTheLoopInputRequestEvent(HumanInTheLoopRequestEvent):
-    """Request free-form text input from a human operator."""
-
-    _display_name: ClassVar[LocaleString] = LocaleString.from_i18n_path("lib.events.hitl_input_request_event.name")
-    _display_description: ClassVar[LocaleString] = LocaleString.from_i18n_path(
-        "lib.events.hitl_input_request_event.description"
-    )
-
-    hitl_type: Annotated[
-        Literal["input"],
-        Field(default="input", description="Fixed to 'input' for text input requests."),
-    ] = "input"
-
-
-class HumanInTheLoopConfirmationRequestEvent(HumanInTheLoopRequestEvent):
-    """Request yes/no confirmation from a human operator."""
-
-    _display_name: ClassVar[LocaleString] = LocaleString.from_i18n_path(
-        "lib.events.hitl_confirmation_request_event.name"
-    )
-    _display_description: ClassVar[LocaleString] = LocaleString.from_i18n_path(
-        "lib.events.hitl_confirmation_request_event.description"
-    )
-
-    hitl_type: Annotated[
-        Literal["confirmation"],
-        Field(default="confirmation", description="Fixed to 'confirmation' for yes/no requests."),
-    ] = "confirmation"
-
-
-class HumanInTheLoopChatRequestEvent(HumanInTheLoopRequestEvent):
-    """Request chat-style input from a human operator.
-
-    Unlike input/confirmation types that show popup dialogs, chat requests appear
-    as regular chat messages. The user responds by typing a normal chat message.
-    """
-
-    _display_name: ClassVar[LocaleString] = LocaleString.from_i18n_path("lib.events.hitl_chat_request_event.name")
-    _display_description: ClassVar[LocaleString] = LocaleString.from_i18n_path(
-        "lib.events.hitl_chat_request_event.description"
-    )
-
-    hitl_type: Annotated[
-        Literal["chat"],
-        Field(default="chat", description="Fixed to 'chat' for chat-style requests."),
-    ] = "chat"
