+++ aihub_lib/aihub_lib/nats/events/human_in_the_loop/request/__init__.py
@@ -1,9 +1,7 @@
-from .HumanInTheLoopRequestEvent import (
-    HumanInTheLoopChatRequestEvent,
-    HumanInTheLoopConfirmationRequestEvent,
-    HumanInTheLoopInputRequestEvent,
-    HumanInTheLoopRequestEvent,
-)
+from .HumanInTheLoopChatRequestEvent import HumanInTheLoopChatRequestEvent
+from .HumanInTheLoopConfirmationRequestEvent import HumanInTheLoopConfirmationRequestEvent
+from .HumanInTheLoopInputRequestEvent import HumanInTheLoopInputRequestEvent
+from .HumanInTheLoopRequestEvent import HumanInTheLoopRequestEvent
 
 __all__ = [
     "HumanInTheLoopChatRequestEvent",
